feat: DecksTable modularization and cleanup
