+++ src/components/tables/decks/index.ts
@@ -1,0 +1,4 @@
+export * from "./DecksTable";
+export * from "./DecksTableColumns";
+export * from "./DecksTableFilters";
+export * from "./DecksTableToolbar";
