+++ src/components/tables/index.ts
@@ -1,3 +1,3 @@
 export * from "./CommanderColors";
-export * from "./DecksTable";
 export * from "./EnhancedTableHeader";
+export * from "./decks";
