+++ src/constants/mtgConstants.ts
@@ -1,0 +1,10 @@
+export const COLOR_MAP: Record<string, string> = {
+  W: "white",
+  U: "blue",
+  B: "black",
+  R: "red",
+  G: "green",
+  C: "colorless",
+};
+
+export const COLOR_ORDER = ["W", "B", "U", "R", "G"];
